// types/components/flickering-grid.ts
import type { HTMLAttributes } from 'react';

export interface FlickeringGridProps extends HTMLAttributes<HTMLDivElement> {
    squareSize?: number;
    gridGap?: number;
    flickerChance?: number;
    color?: string;
    width?: number;
    height?: number;
    className?: string;
    maxOpacity?: number;
}